Add files via upload
It is well known that cheap stocks have continued to under perform the broader market for some time now. The question is do we need to re-evaluate our definition of cheap or "value" stocks. This analysis shows that some valuation metrics, such as price to book, may no longer be viable in determining a stock's true valuation. As, intellectual property continues to increase, book value may be less of reliable metric.